+++ service_catalog/forms/__init__.py
@@ -11,3 +11,4 @@ from .service_request_forms import *
 from .support_message_forms import *
 from .support_request_forms import *
 from .tower_server_forms import *
+from .form_utils import *